requirements.txt
crewai
crewai_tools
python-dotenv
langchain
replicate
langchain-community
duckduckgo-search